% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fcomplete.R
\name{plot.fcomplete}
\alias{plot.fcomplete}
\title{Plots the fcomplete fit object}
\usage{
\method{plot}{fcomplete}(model, rows = 1, true = NULL, test = NULL,
  title = NULL, filename = NULL, ylim = NULL, width = 10,
  height = 10)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of dimensions}

\item{K}{true 'low dimension'}

\item{dgrid}{size of the grid}

\item{clear}{fraction of observations to remove}

\item{noise_mag}{the magnitude of noise}
}
\description{
Plots the fcomplete fit object
}
\details{
The function simulates data as described in the paper.
Generates observations \code{X1,X2} and a \code{Y = X1 + X2 + noise}
}
\seealso{
\code{\link{fregression}}
}
