% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsimulate.R
\name{fsimulate}
\alias{fsimulate}
\title{Simulate data}
\usage{
fsimulate(n = 100, d = 7, K = 3, dgrid = 51, clear = 0.85,
  num_points = NULL, noise.mag = 0.1)
}
\arguments{
\item{n}{number of observations}

\item{d}{number of dimensions}

\item{K}{true 'low dimension'}

\item{dgrid}{size of the grid}

\item{clear}{fraction of observations to remove}

\item{num_points}{number of points observed per curve}

\item{noise_mag}{the magnitude of noise}
}
\description{
Function simulates data from the paper
}
\details{
The function simulates data as described in the paper.
Generates observations \code{X1,X2} and a \code{Y = X1 + X2 + noise}
}
\seealso{
\code{\link{fregression}}
}
